backend/requirements.txt
uvicorn
fastapi
python-dotenv
pydantic
sqlalchemy
pymysql
passlib
cryptography